Move StartParameterMutationReportingSettingsProcessor to internal package
fixes sanityCheck
